(PC-28265)[PRO] refactor: simplify FormLayoutSection
