+++ pro/src/components/FormLayout/FormLayoutSection.tsx
@@ -2,19 +2,15 @@ import cn from 'classnames'
 import React from 'react'
 
 import { Title } from 'ui-kit'
-import { Link } from 'ui-kit/Banners/LinkNodes/LinkNodes'
 
 import style from './FormLayout.module.scss'
-
-import { FormLayoutDescription } from '.'
+import { FormLayoutDescription } from './FormLayoutDescription'
 
 interface FormLayoutSectionProps {
   title: React.ReactNode
   description?: string
   children: React.ReactNode | React.ReactNode[]
   className?: string
-  descriptionAsBanner?: boolean
-  links?: Link[]
   id?: string
 }
 
@@ -23,8 +19,6 @@ const Section = ({
   description,
   children,
   className,
-  descriptionAsBanner = false,
-  links,
   id,
 }: FormLayoutSectionProps): JSX.Element => (
   <fieldset className={cn(style['form-layout-section'], className)} id={id}>
@@ -34,11 +28,7 @@ const Section = ({
       </Title>
     </legend>
     <div className={style['form-layout-section-header']}>
-      <FormLayoutDescription
-        description={description}
-        isBanner={descriptionAsBanner}
-        links={links}
-      />
+      <FormLayoutDescription description={description} />
     </div>
     {children}
   </fieldset>
