+++ pro/src/components/FormLayout/index.tsx
@@ -1,4 +1,1 @@
-import FormLayoutDescription from './FormLayoutDescription'
-
 export { default } from './FormLayout'
-export { FormLayoutDescription }
