+++ pro/src/screens/IndividualOffer/StocksThing/StocksThing.tsx
@@ -9,7 +9,8 @@ import {
   GetOfferStockResponseModel,
   SubcategoryIdEnum,
 } from 'apiClient/v1'
-import FormLayout, { FormLayoutDescription } from 'components/FormLayout'
+import FormLayout from 'components/FormLayout'
+import { FormLayoutDescription } from 'components/FormLayout/FormLayoutDescription'
 import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferNavigation/constants'
 import { RouteLeavingGuardIndividualOffer } from 'components/RouteLeavingGuardIndividualOffer/RouteLeavingGuardIndividualOffer'
 import { StockFormActions } from 'components/StockFormActions'
